mawk - (Ver. 1.1) An interpreter for the AWK language
Mawk is an interpreter for the AWK programming language.

Mawk 1.0 implemented the language as defined in the 1988 AWK book.
Version 1.1 adds new features:

     -- Full posix1003.2 draft(11.2) compliance

	toupper(), tolower(), CONVFMT and ENVIRON[]

	-v var=value command line assignments

	Ansi C compatible printf() and sprintf()

     -- Hard coded limits are either removed or much larger

     -- MsDOS added, system() and pipes.  BINMODE variable to
	control CRLF translation.

Mawk is distributed without warranty under the terms of the
GNU General Public License, version 2, June 1991.  The file
COPYING contains the GNU General Public License, version 2.

Mawk 1.1 has been test on:

     sunos4.1.1  sun3 and sun4   (cc and gcc)
     sunos4.0.3  sun3 and sun4   (cc and gcc)
     ultrix4.2   decstation 5000 (mips) (cc and gcc)
     ultrix4.1   microvax 3600
     ultrix3.1   vax
     bsd43	 vax
     SysVR3	 SCO3.2.2      (cc, rcc and gcc)
     ESIX System 5.3.2 RevD using gcc 1.39
     SysVR3      Stardent 3000
     MsDOS	 TurboC++ 1.0 and BCC++, MSC5.1 and 6.0
     MIPS M2000 	C 2.20 (4.52)
     Atarist	 cross compile with gcc

Mawk 1.0 compiled and tested on the following which should still
work with at most minor tweaking.

	XENIX SysV R2.3.1
	Dynix on Sequent Symmetry
	apolloSR10.3

mawk1.09 a beta release of mawk1.1 can be obtained by
anon ftp from oxy.edu.

Mawk is an interpreter for the AWK programming language.
Mawk 1.0 implemented the language as defined in the 1988 AWK book.
This release adds features required by posix1003.2 draft 11.2:

    multiple -f options and -v var=value option
    toupper(), tolower()
    ENVIRON[] and CONVFMT
    ANSI C compatible printf() and sprintf()

Hard coded limits on the number of fields is gone, and system()
and pipes have been added for MsDOS.

The manual is not up to date, and testing on more than the
systems I have access to are needed to bring this release to 1.1.