Serious bug in mawk 1.3.2
There is a serious bug in mawk1.3.2. Some regular expression character
classes aren't handled correctly.

This bug is only in 1.3.2 and not in previous releases.

Very few awk programs actually expose the bug.

The bug is fixed in release 1.3.3.

You can ftp 1.3.3 from ftp://ftp.whidbey.net/pub/brennan.

As the bug was a 1 character typo, the diffs 1.3.2 -> 1.3.3 are small
and appended below.